Fixed circleci issue, but tests wont run for booking

+++ BackEnd/src/main/java/septmajorproject/bookingsys/model/Customer.java
@@ -115,7 +115,7 @@ public class Customer {
         this.updatedDate = updatedDate;
     }
 
-    public long getId()
+    public Integer getId()
     {
         return id;
     }
